finish renaming gem
