+++ lib/fast-jsonapi.rb
